[WIP] Let the client application continue execution after failed decode
Sometimes the IB api returns invalid data. This raises a ValueError in
decoder. The error is catched by the decoder and the event loop
continues. In the synchronous mode this means that the client
application halts and never resumes execution.

The fix in this commit calls the request end handler after the exception
is logged.